Add AbstractKey.SwitchLanguage
