+++ Assets/Nova/Sources/Core/Input/AbstractKey.cs
@@ -17,6 +17,7 @@ namespace Nova
         ReturnTitle,
         QuitGame,
         ToggleFullScreen,
+        SwitchLanguage,
         LeaveView,
         EditorBackward,
         EditorBeginChapter,
