--- Assets/Nova/Sources/Scripts/ToggleFullScreen.cs
@@ -1,22 +1,0 @@
-﻿using UnityEngine;
-
-namespace Nova
-{
-    public class ToggleFullScreen : MonoBehaviour
-    {
-        private InputManager inputManager;
-
-        private void Awake()
-        {
-            inputManager = Utils.FindNovaGameController().InputManager;
-        }
-
-        private void Update()
-        {
-            if (inputManager.IsTriggered(AbstractKey.ToggleFullScreen))
-            {
-                GameRenderManager.SwitchFullScreen();
-            }
-        }
-    }
-}
