+++ Assets/Scenes/Main.unity
@@ -371,7 +371,6 @@ GameObject:
   m_Component:
   - component: {fileID: 339668545}
   - component: {fileID: 339668544}
-  - component: {fileID: 339668546}
   - component: {fileID: 339668548}
   - component: {fileID: 339668547}
   - component: {fileID: 339668549}
@@ -412,18 +411,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &339668546
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 339668543}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9e6baf31cdb4cc24092d1fb595637905, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &339668547
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1212,6 +1199,7 @@ GameObject:
   - component: {fileID: 686561426}
   - component: {fileID: 686561425}
   - component: {fileID: 686561428}
+  - component: {fileID: 686561429}
   m_Layer: 0
   m_Name: EventSystem
   m_TagString: Untagged
@@ -1287,6 +1275,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 3eafaa8a04f585446953da5981667e0b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &686561429
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 686561424}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9e6baf31cdb4cc24092d1fb595637905, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &702255950
